third_party/WebKit/LayoutTests/fast/dom/shadow/content-combinator-with-shadow-deep-expected.txt

Test for /content/ with /shadow-deep/ and :host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorOf("#host1 > .content") is "rgb(0, 128, 0)"
PASS colorOf("#host1 > .content") is "rgb(0, 128, 0)"
PASS colorOf("#host1 > .content") is "rgb(0, 128, 0)"
PASS colorOf("#host1 > .content") is not "rgb(255, 0, 0)"
PASS colorOf("#host1 > .content") is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
